库间调拨 - 代码微调

+++ web/Application/Home/Service/ITBillService.class.php
@@ -339,6 +339,9 @@ class ITBillService extends PSIBaseService {
 		return $this->ok($id);
 	}
 
+	/**
+	 * 查询某个调拨单的详情
+	 */
 	public function itBillInfo($params) {
 		if ($this->isNotOnline()) {
 			return $this->emptyResult();
@@ -401,6 +404,9 @@ class ITBillService extends PSIBaseService {
 		return $result;
 	}
 
+	/**
+	 * 调拨单的明细记录
+	 */
 	public function itBillDetailList($params) {
 		if ($this->isNotOnline()) {
 			return $this->emptyResult();
@@ -429,6 +435,9 @@ class ITBillService extends PSIBaseService {
 		return $result;
 	}
 
+	/**
+	 * 删除调拨单
+	 */
 	public function deleteITBill($params) {
 		if ($this->isNotOnline()) {
 			return $this->notOnlineError();
@@ -437,10 +446,13 @@ class ITBillService extends PSIBaseService {
 		$id = $params["id"];
 		
 		$db = M();
+		$db->startTrans();
+		
 		$sql = "select ref, bill_status from t_it_bill where id = '%s' ";
 		$data = $db->query($sql, $id);
 		
 		if (! $data) {
+			$db->rollback();
 			return $this->bad("要删除的调拨单不存在");
 		}
 		
@@ -448,26 +460,29 @@ class ITBillService extends PSIBaseService {
 		$billStatus = $data[0]["bill_status"];
 		
 		if ($billStatus != 0) {
+			$db->rollback();
 			return $this->bad("调拨单(单号：$ref)已经提交，不能被删除");
 		}
 		
-		$db->startTrans();
-		try {
-			$sql = "delete from t_it_bill_detail where itbill_id = '%s' ";
-			$db->execute($sql, $id);
-			
-			$sql = "delete from t_it_bill where id = '%s' ";
-			$db->execute($sql, $id);
-			
-			$bs = new BizlogService();
-			$log = "删除调拨单，单号：$ref";
-			$bs->insertBizlog($log, "库间调拨");
-			
-			$db->commit();
-		} catch ( Exception $e ) {
+		$sql = "delete from t_it_bill_detail where itbill_id = '%s' ";
+		$rc = $db->execute($sql, $id);
+		if ($rc === false) {
 			$db->rollback();
-			return $this->bad("数据库错误，请联系系统管理员");
+			return $this->sqlError(__LINE__);
 		}
+		
+		$sql = "delete from t_it_bill where id = '%s' ";
+		$rc = $db->execute($sql, $id);
+		if ($rc === false) {
+			$db->rollback();
+			return $this->sqlError(__LINE__);
+		}
+		
+		$bs = new BizlogService();
+		$log = "删除调拨单，单号：$ref";
+		$bs->insertBizlog($log, $this->LOG_CATEGORY);
+		
+		$db->commit();
 		
 		return $this->ok();
 	}
